backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15,<4.0
django-cors-headers>=4.3,<5.0
gunicorn>=22.0,<23.0
openai>=1.0.0
google-generativeai>=0.3.0
whitenoise>=6.6.0
